[REFACTOR] store 파일명 변경
