+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
 import { Provider } from 'react-redux';
-import { store } from '@/store/ingredientStore';
+import { store } from '@/store';
 import '@/styles/globals.css';
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
